Make several "Delete" options
"Delete" option made when doing long press for some cardviews:

Incomes
Transactions
