+++ app/src/main/res/menu/income_item_menu.xml
@@ -1,11 +1,11 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/menu_item_edit"
+        android:id="@+id/menu_item_edit_income"
         android:title="@string/menu_edit"
         app:showAsAction="never"/>
     <item
-        android:id="@+id/menu_item_delete"
+        android:id="@+id/menu_item_delete_income"
         android:title="@string/menu_delete"
         app:showAsAction="never"/>
 </menu>
